requirements.txt
pip==22.3.1
setuptools==65.5.1
wheel==0.38.4
pytest
playwright
allure-pytest
requests
pytest-playwright-snapshot==1.0
axe-playwright-python==0.1.4
